# docs/openapi.yaml
openapi: 3.0.0
info:
  title: LLM Evaluation Pipeline API
  description: |
    Comprehensive LLM evaluation framework for Turkish and fintech-focused model testing.
    
    ## Features
    - Turkish language benchmarks (grammar, creativity, reasoning, nuance)
    - Fintech domain testing (knowledge, calculations, compliance)
    - Function calling and agentic workflows
    - Multi-turn conversations and RAG evaluation
    - PII detection and safety testing
    - Adversarial resilience and jailbreak testing
    - Language mix handling (Turkish-English)
    - Self-consistency and prompt compression
    - Human-in-the-loop feedback integration
    
  version: 2.0.0
  contact:
    name: LLM Eval Pipeline Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8501
    description: Local Streamlit Dashboard
  - url: http://localhost:8000
    description: Local API Server (if deployed)
...
paths:
...
  /api/v1/reports:
    get:
      tags:
        - Reports
      summary: List available reports
      description: Get list of all generated evaluation reports
      operationId: listReports
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Reports list retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportsList'
  /api/v1/reports/{reportId}:
    get:
      tags:
        - Reports
      summary: Get specific report
      description: Retrieve detailed evaluation report by ID
      operationId: getReport
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Report retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationReport'
        '404':
          description: Report not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/reports/{reportId}/compare:
    post:
      tags:
        - Reports
      summary: Compare multiple reports
      description: Generate comparison between multiple evaluation reports
      operationId: compareReports
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                compare_with:
                  type: array
                  items:
                    type: string
                  description: List of report IDs to compare with
      responses:
        '200':
          description: Comparison generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComparisonReport'
...
components:
  schemas:
...
    EvaluationSummary:
      type: object
      properties:
        total_tests:
          type: integer
        total_cost:
          type: number
        total_time:
          type: number
        model_comparison:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ModelComparison'
        best_performers:
          type: object
    ModelComparison:
      type: object
      properties:
        overall_score:
          type: number
          minimum: 0
          maximum: 1
        avg_latency:
          type: number
        latency_p95:
          type: number
        total_cost:
          type: number
        passed_thresholds:
          type: boolean
    ModelResults:
      type: object
      properties:
        model_name:
          type: string
        tests:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/TestResults'
    TestResults:
      type: object
      properties:
        summary:
          $ref: '#/components/schemas/TestSummary'
        results:
          type: array
          items:
            type: object
    TestSummary:
      type: object
      properties:
        overall_score:
          type: number
        total_tests:
          type: integer
        avg_latency:
          type: number
        total_cost:
          type: number
        avg_scores:
          type: object
          additionalProperties:
            type: number
...
    ReportsList:
      type: object
      properties:
        total:
          type: integer
        reports:
          type: array
          items:
            $ref: '#/components/schemas/ReportSummary'
    ReportSummary:
      type: object
      properties:
        report_id:
          type: string
        timestamp:
          type: string
          format: date-time
        models:
          type: array
          items:
            type: string
        test_suite:
          type: string
        overall_score:
          type: number
    EvaluationReport:
      type: object
      properties:
        report_id:
          type: string
        timestamp:
          type: string
        summary:
          $ref: '#/components/schemas/EvaluationSummary'
        models:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ModelResults'
    ComparisonReport:
      type: object
      properties:
        reports:
          type: array
          items:
            type: string
        comparison:
          type: object
          properties:
            score_changes:
              type: object
            performance_trends:
              type: object
            cost_analysis:
              type: object
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object